Unifico los modales de manejo de sesión y los linkeo con el header
