+++ src/components/Header/Header.js
@@ -11,16 +11,18 @@ import MenuIcon from '@mui/icons-material/Menu';
 import { useNavigate } from 'react-router-dom';
 import airbnbLogo from '../../assets/airbnbLogo.svg';
 import { headerStyles } from './HeaderStyles';
+import SesionModal from '../SesionModal/SesionModal';
 
 const Header = () => {
   const styles = headerStyles();
   const theme = useTheme();
   const [anchorMenu, setAnchorMenu] = useState();
+  const [openLoginModal, setOpenLoginModal] = useState(false);
   const openMenu = !!anchorMenu;
   const navigate = useNavigate();
 
   // Usuario mockeado---------
-  const [userIsLogged, setUserIsLogged] = useState(true);
+  const [userIsLogged, setUserIsLogged] = useState(false);
   const user = {
     firstName: 'John',
     lastName: 'Doue',
@@ -33,9 +35,7 @@ const Header = () => {
     if (userIsLogged) {
       setAnchorMenu(currentTarget);
     } else {
-      // Acá va la función que hace que se abra el modal del login
-      // Seguramente tengamos que manejar por context el login
-      setUserIsLogged(true);
+      setOpenLoginModal(true);
     }
   };
 
@@ -46,6 +46,10 @@ const Header = () => {
   const handleLogOut = () => {
     setUserIsLogged(false);
     handleCloseMenu();
+  };
+
+  const onCloseLoginModal = () => {
+    setOpenLoginModal(false);
   };
 
   return (
@@ -69,6 +73,7 @@ const Header = () => {
         <Divider />
         <MenuItem type="text" onClick={handleLogOut}>Cerrar Sesión</MenuItem>
       </Menu>
+      <SesionModal open={openLoginModal} onClose={onCloseLoginModal} />
     </>
   );
 };
